updated: ChatGPT组件v1.1.0
