+++ packages/tdp-editor-components/package.json
@@ -19,8 +19,6 @@
     "dist"
   ],
   "dependencies": {
-    "@ant-design/icons-vue": "^6.1.0",
-    "axios": "^0.24.0",
     "classnames": "^2.3.1",
     "moment": "^2.29.1",
     "openai": "^3.2.1",
